Rename fields.scss to editor-base.scss
