+++ ts/editor/OnlyEditable.svelte
@@ -4,7 +4,7 @@ License: GNU AGPL, version 3 or later; http://www.gnu.org/licenses/agpl.html
 -->
 <script lang="typescript">
     import WithContext from "components/WithContext.svelte";
-    import { fieldFocusedKey, inCodableKey } from "./context-keys";
+    import { fieldFocusedKey, inCodableKey } from "./NoteEditor.svelte";
 </script>
 
 <WithContext key={fieldFocusedKey} let:context={fieldFocused}>
