+++ ts/editor/TemplateButtons.svelte
@@ -5,7 +5,7 @@ License: GNU AGPL, version 3 or later; http://www.gnu.org/licenses/agpl.html
 <script lang="typescript">
     import * as tr from "lib/i18n";
     import { bridgeCommand } from "lib/bridgecommand";
-    import { fieldFocusedKey, inCodableKey } from "./context-keys";
+    import { fieldFocusedKey, inCodableKey } from "./NoteEditor.svelte";
 
     import ButtonGroup from "components/ButtonGroup.svelte";
     import ButtonGroupItem from "components/ButtonGroupItem.svelte";
