--- ts/editor/context-keys.ts
@@ -1,5 +1,0 @@
-// Copyright: Ankitects Pty Ltd and contributors
-// License: GNU AGPL, version 3 or later; http://www.gnu.org/licenses/agpl.html
-
-export const fieldFocusedKey = Symbol("fieldFocused");
-export const inCodableKey = Symbol("inCodable");
