+++ ts/editor/editor-base.scss
@@ -43,6 +43,7 @@ body {
     }
 }
 
+/* CodeMirror */
 @import "codemirror/lib/codemirror";
 @import "codemirror/theme/monokai";
 @import "codemirror/addon/fold/foldgutter";
